backend/requirements.txt
fastapi==0.115.13           
pydantic>=2.11.1,<3.0.0  
pydantic-settings==2.10.0   
uvicorn[standard]==0.23.2   
SQLAlchemy==2.0.41          
asyncpg==0.28.0             
alembic==1.16.2             
python-dotenv==1.1.0        
passlib==1.7.4
pydantic[email]==2.11.1
python-dateutil==2.9.0.post0
six==1.17.0
python-jose==3.5.0
python-multipart==0.0.20